feat: separate message component
